+++ waitlst/src/views/AuthView.vue
@@ -1,69 +1,20 @@
 <script setup lang="ts">
-import { Button } from '@/components/ui/button'
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
-import { Input } from '@/components/ui/input'
-import { Label } from '@/components/ui/label'
-import { supabase } from '@/utils/supabase.ts'
 import { useAuthStore } from '@/stores/AuthStore.ts'
 import { useRouter } from 'vue-router'
-
-import { PulseLoader } from '@/components/ui/pulse-loader'
-import { onMounted, ref, watchEffect } from 'vue'
-
-const isLoading = ref(false)
-const email = ref('')
+import { onMounted, watchEffect } from 'vue'
+import EmailLoginForm from '@/components/AuthEmailLoginForm.vue'
+import OAuthLoginButton from '@/components/AuthGitHubLogin.vue'
 
 const router = useRouter()
 const authStore = useAuthStore()
 
-const handleEmailLogin = async () => {
-  try {
-    isLoading.value = true
-    const { error } = await supabase.auth.signInWithOtp({
-      email: email.value,
-      options: {
-        emailRedirectTo: `${window.location.origin}/dashboard`,
-      },
-    })
-    if (error) throw error
-  } catch (error) {
-    if (error instanceof Error) {
-      alert(error.message)
-    }
-  } finally {
-    isLoading.value = false
-  }
-}
-
-const signInWithGithub = async () => {
-  try {
-    isLoading.value = true
-    const { error } = await supabase.auth.signInWithOAuth({
-      provider: 'github',
-      options: {
-        redirectTo: `${window.location.origin}/dashboard`,
-      },
-    })
-    if (error) throw error
-  } catch (error) {
-    if (error instanceof Error) {
-      alert(error.message)
-    }
-  } finally {
-    isLoading.value = false
-  }
-}
-
-onMounted(async () => {
-  if (authStore.id) {
-    router.push('/dashboard')
-  }
+onMounted(() => {
+  if (authStore.id) router.push('/dashboard')
 })
 
 watchEffect(() => {
-  if (authStore.id) {
-    router.push('/dashboard')
-  }
+  if (authStore.id) router.push('/dashboard')
 })
 </script>
 
@@ -75,39 +26,11 @@ watchEffect(() => {
         <CardDescription>If you already have an account, we'll log you in. </CardDescription>
       </CardHeader>
       <CardContent>
-        <form @submit.prevent="handleEmailLogin">
-          <div class="grid gap-4">
-            <div class="grid gap-2">
-              <Label for="email">Email</Label>
-              <Input
-                id="email"
-                placeholder="name@example.com"
-                type="email"
-                auto-capitalize="none"
-                auto-complete="email"
-                auto-correct="off"
-                :disabled="isLoading"
-                v-model="email"
-              />
-            </div>
-            <Button type="submit" :disabled="isLoading">
-              <template v-if="isLoading">
-                <PulseLoader :loading="true" size="8px" margin="5px" color-class="bg-black" />
-              </template>
-              <template v-else> Sign In with Email </template>
-            </Button>
-            <div class="h-0.5 w-full bg-secondary"></div>
-            <Button type="button" @click="signInWithGithub" :disabled="isLoading" variant="outline">
-              <template v-if="isLoading">
-                <PulseLoader :loading="true" size="8px" margin="5px" color-class="bg-white" />
-              </template>
-              <template v-else>
-                <img src="@/assets/github-mark-white.svg" alt="GitHub Logo" class="h-4 w-4 mr-2" />
-                Continue with GitHub
-              </template>
-            </Button>
-          </div>
-        </form>
+        <div class="flex flex-col gap-y-4">
+          <EmailLoginForm />
+          <div class="h-0.5 w-full bg-secondary"></div>
+          <OAuthLoginButton />
+        </div>
       </CardContent>
     </Card>
   </div>
